Fixed NumberOfShotsBug

+++ code/jvm/daw-battleship-game/src/main/kotlin/pt/isel/daw/dawbattleshipgame/domain/game/Configuration.kt
@@ -16,7 +16,7 @@ data class Configuration(
         @JsonProperty("fleet")
     val fleet: Map<ShipType, Int>, // List<Ship:Occupation>
         @JsonProperty("shots")
-    val shots: Int,
+    val shots: Long,
         @JsonProperty("roundTimeout")
     val roundTimeout: Long
 ) {
